Fixed: Simplified code
- Output has to be formatted...still ;)
