+++ shared.py
@@ -3,10 +3,8 @@ import paramiko
 import time
 import subprocess
 import json
-import datetime
-import matplotlib.pyplot as plt
 
-yourPw = "yourpasswd"
+PASSWD = "yourpasswd"
 
 def prepare_mapago(ctx):
     print("Preparing mapago")
@@ -21,7 +19,6 @@ def netem_reset(ctx, hostname, interfaces):
 
 
 def netem_configure(ctx, hostname, interfaces=[], netem_params={'rate' : '100kbit'}):
-    supported_params = ['rate', 'loss', 'delay']
     qdisc_id = 10
     netem_used = False
   
@@ -33,17 +30,14 @@ def netem_configure(ctx, hostname, interfaces=[], netem_params={'rate' : '100kbi
 
         # 2. sweep through desired params
         for param in netem_params:
-
-            # 3. check if param is in supported netem_params
-            if any(param in s for s in supported_params):
-                if param == 'rate':
-                    continue
-                elif param == 'loss':
-                    netem_used = True
-                    netem_cmd += " loss {}%".format(netem_params[param])
-                elif param == 'delay':
-                    netem_used = True
-                    netem_cmd += " delay {}ms".format(netem_params[param])
+            if param == 'rate':
+                continue
+            elif param == 'loss':
+                netem_used = True
+                netem_cmd += " loss {}%".format(netem_params[param])
+            elif param == 'delay':
+                netem_used = True
+                netem_cmd += " delay {}ms".format(netem_params[param])
                 ### check for future params ###
             else:
                 raise Exception('\nParam {} not supported!'.format(param))
@@ -128,92 +122,6 @@ def prepare_client(ctx, params):
 
     return msmt_db
 
-# copy & paste from evaluation.py
-def analyze_data(msmt_results):
-    print("\nDUMMY: Calling matplotlib to analyze msmt results")
-    print("\nmsmt_res are: {}".format(msmt_results))
-
-    # min and max will at the end of the next loop contain
-    # the real min and max values
-    datetime_min = datetime.datetime(4000, 1, 1)
-    datetime_max = datetime.datetime(1, 1, 1)
-
-    # Keep in mind for all byte accounting here:
-    # it is transmitted in a ABSOLUTE manner, no differences
-    # are signaled via mapago
-    bytes_rx = 0
-    prev_datetime_max = datetime.datetime(1, 1, 1)
-    prev_datetime_max_set = False
-
-    # we store the data for plot, where just one overall
-    # bandwith is not succiently,
-    normalized = []
-
-    # now find the youngest and oldest date
-    for entry in msmt_results:
-        bytes_measurement_point = 0
-
-        # one entry can have multiple streams, so iterate over the
-        # streams now
-        for stream in entry:
-            time = datetime.datetime.strptime(
-                stream['ts-start'], '%Y-%m-%dT%H:%M:%S.%f')
-        
-            if time < datetime_min:
-                datetime_min = time
-            
-                if prev_datetime_max_set == False:
-                    prev_datetime_max = datetime_min
-                    prev_datetime_max_set = True 
-
-            time = datetime.datetime.strptime(
-                stream['ts-end'], '%Y-%m-%dT%H:%M:%S.%f')
-        
-            if time > datetime_max:
-                datetime_max = time
-
-            bytes_measurement_point += int(stream['bytes'])
-
-        # we got data from all streams of that entry
-        curr_msmt_time = (datetime_max - datetime_min).total_seconds()
-        bytes_per_period = bytes_measurement_point - bytes_rx
-        mbits_per_period = (bytes_per_period * 8) / 10**6
-        # bytes_rx == # bytes until now received
-        bytes_rx = bytes_measurement_point
-
-        # this works only if data is send immediately after prev_datetime_max
-        # or we pay attention to a period where nothing is transmitted
-        duration_of_period = (datetime_max - prev_datetime_max).total_seconds()
-        prev_datetime_max = datetime_max
-        throughput_of_period = mbits_per_period / duration_of_period
-        normalized.append([curr_msmt_time, throughput_of_period])
-
-    measurement_length = (datetime_max - datetime_min).total_seconds()
-    bytes_sec = bytes_rx / measurement_length
-    Mbits_sec = (bytes_sec * 8) / 10**6
-    Kbits_sec = (bytes_sec * 8) / 10**3
-    print('overall bandwith: {} bytes/sec'.format(bytes_sec))
-    print('overall bandwith: {} Mbits/sec'.format(Mbits_sec))
-    print('overall bandwith: {} Kbits/sec'.format(Kbits_sec))
-    print('measurement length: {} sec]'.format(measurement_length))
-    print('received: {} bytes]'.format(bytes_rx))
-
-    # now plotting starts, not really fancy
-
-    # normalize date to start with just 0 sec and not 2018-01-23 ...
-    x = []
-    y = []
-    for i in normalized:
-        x.append(i[0])
-        y.append(i[1])
-
-    fig = plt.figure()
-    plt.plot(x, y)
-    plt.ylabel('Throughput [MBits/s]')
-    plt.xlabel('Time [seconds]')
-    fig.savefig("./data/msmtResult.pdf", bbox_inches='tight')
-   
-
 def host_alive(ctx, hostname):
     print("\nChecking host {} with addr {}".format(hostname, ctx.config[hostname]['ip-ctrl']))
 
@@ -250,7 +158,7 @@ def ssh_execute(ctx, hostname, command, timeout=10, background=False):
         # ...still we get an authentication error
         # ssh.connect(host, int(port), username, password, pkey=pkey, timeout=1.0)
         # use temporarily user/pwd not certificate
-        ssh.connect(host, int(port), username, yourPw)
+        ssh.connect(host, int(port), username, PASSWD)
 
         if background:
             transport = ssh.get_transport()
